+++ main/src/main/scala/org/clulab/processors/Sentence.scala
@@ -11,13 +11,21 @@ import scala.util.hashing.MurmurHash3._
 
 /** Stores the annotations for a single sentence */
 class Sentence(
-  /** Actual tokens in this sentence */
-  val words: Array[String],
-  /** Start character offsets for the words; start at 0 */
+  /** Raw tokens in this sentence; these MUST match the original text */
+  val raw: Array[String],
+  /** Start character offsets for the raw tokens; start at 0 */
   val startOffsets: Array[Int],
-  /** End character offsets for the words; start at 0 */
-  val endOffsets: Array[Int]
-) extends Serializable {
+  /** End character offsets for the raw tokens; start at 0 */
+  val endOffsets: Array[Int],
+
+  /**
+    * Words produced from raw tokens, closer to what the downstream components expect
+    * These MAY differ from raw tokens,
+    *   e.g., Unicode characters in raw are replaced with ASCII strings, and parens are replaced with -LRB-, -RRB-, etc.
+    * However, the number of raw tokens MUST always equal the number of words, so if the exact text must be recovered,
+    *   please use the raw tokens with the same positions
+    */
+  val words: Array[String]) extends Serializable {
 
   /** POS tags for words */
   var tags: Option[Array[String]] = None
@@ -34,7 +42,7 @@ class Sentence(
   /** DAG of syntactic and semantic dependencies; word offsets start at 0 */
   var graphs: GraphMap = new GraphMap
 
-  def size:Int = words.length
+  def size:Int = raw.length
 
   def indices: Range = 0 until size
 
@@ -59,8 +67,9 @@ class Sentence(
     // decided to use the class name
     val h0 = stringHash(stringCode)
     // NOTE: words.hashCode will produce inconsistent values
-    val h1 = mix(h0, getAnnotationsHash(Some(words)))
-    val h2 = mix(h1, getAnnotationsHash(Some(startOffsets)))
+    val h1a = mix(h0, getAnnotationsHash(Some(raw)))
+    val h1b = mix(h1a, getAnnotationsHash(Some(words)))
+    val h2 = mix(h1b, getAnnotationsHash(Some(startOffsets)))
     val h3 = mix(h2, getAnnotationsHash(Some(endOffsets)))
     val h4 = mix(h3, getAnnotationsHash(tags))
     val h5 = mix(h4, getAnnotationsHash(lemmas))
@@ -68,7 +77,7 @@ class Sentence(
     val h7 = mix(h6, getAnnotationsHash(norms))
     val h8 = mix(h7, getAnnotationsHash(chunks))
     val h9 = mix(h8, if (dependencies.nonEmpty) dependencies.get.equivalenceHash else None.hashCode)
-    finalizeHash(h9, 9)
+    finalizeHash(h9, 9) // TODO: is 9 still valid, now that we have raw tokens as well?
   }
 
   /**
@@ -103,24 +112,22 @@ class Sentence(
     *
     * @return the text of the sentence
     */
-  def getSentenceText():String =  getSentenceFragmentText(0, words.length)
+  def getSentenceText:String =  getSentenceFragmentText(0, words.length)
 
   def getSentenceFragmentText(start:Int, end:Int):String = {
     // optimize the single token case
-    if(end - start == 1) words(start)
+    if(end - start == 1) return raw(start)
 
     val text = new mutable.StringBuilder()
     for(i <- start until end) {
       if(i > start) {
         // add as many white spaces as recorded between tokens
-        // sometimes this space is negative: in BioNLPProcessor we replace "/" with "and"
-        //   in these cases, let's make sure we print 1 space, otherwise the text is hard to read
         val numberOfSpaces = math.max(1, startOffsets(i) - endOffsets(i - 1))
         for (j <- 0 until numberOfSpaces) {
           text.append(" ")
         }
       }
-      text.append(words(i))
+      text.append(raw(i))
     }
     text.toString()
   }
@@ -128,9 +135,10 @@ class Sentence(
   /** Reverts the current sentence */
   def revert():Sentence = {
     val reverted = new Sentence(
-      SeqUtils.revert(words).toArray,
+      SeqUtils.revert(raw).toArray,
       SeqUtils.revert(startOffsets).toArray,
-      SeqUtils.revert(endOffsets).toArray)
+      SeqUtils.revert(endOffsets).toArray,
+      SeqUtils.revert(words).toArray)
     if(tags.nonEmpty)
       reverted.tags = Some(SeqUtils.revert(tags.get).toArray)
     if(lemmas.nonEmpty)
@@ -152,14 +160,23 @@ class Sentence(
 object Sentence {
 
   def apply(
-    words: Array[String],
+    raw:Array[String],
     startOffsets: Array[Int],
-    endOffsets: Array[Int]
-  ): Sentence = new Sentence(words, startOffsets, endOffsets)
+    endOffsets: Array[Int]): Sentence =
+    new Sentence(raw, startOffsets, endOffsets, raw) // words are identical to raw tokens (a common situation)
+
   def apply(
-    words: Array[String],
+    raw:Array[String],
     startOffsets: Array[Int],
     endOffsets: Array[Int],
+    words: Array[String]): Sentence =
+    new Sentence(raw, startOffsets, endOffsets, words)
+
+  def apply(
+    raw: Array[String],
+    startOffsets: Array[Int],
+    endOffsets: Array[Int],
+    words: Array[String],
     tags: Option[Array[String]],
     lemmas: Option[Array[String]],
     entities: Option[Array[String]],
@@ -168,7 +185,7 @@ object Sentence {
     tree: Option[Tree],
     deps: GraphMap
   ): Sentence = {
-    val s = Sentence(words, startOffsets, endOffsets)
+    val s = Sentence(raw, startOffsets, endOffsets, words)
     // update annotations
     s.tags = tags
     s.lemmas = lemmas