+++ main/src/main/scala/org/clulab/processors/clu/CluProcessor.scala
@@ -1,10 +1,8 @@
 package org.clulab.processors.clu
 
-import java.io.StringReader
-
 import org.clulab.processors.clu.sequences.{Chunker, NamedEntityRecognizer, PartOfSpeechTagger}
 import org.clulab.processors.clu.syntax._
-import org.clulab.processors.clu.tokenizer.{OpenDomainEnglishTokenizer, Tokenizer}
+import org.clulab.processors.clu.tokenizer._
 import org.clulab.processors.{Document, Processor, Sentence}
 import org.clulab.struct.GraphMap
 import com.typesafe.config.{Config, ConfigFactory}
@@ -16,9 +14,6 @@ import scala.collection.mutable
 import scala.collection.mutable.ArrayBuffer
 import CluProcessor._
 import org.clulab.sequences.{LexiconNER, Tagger}
-import uk.ac.susx.informatics.Morpha
-
-import scala.util.matching.Regex
 
 /**
   * Processor that uses only tools that are under Apache License
@@ -35,22 +30,8 @@ class CluProcessor (val config: Config = ConfigFactory.load("cluprocessoropen"))
   // should we intern strings or not?
   val internStrings:Boolean = getArgBoolean(s"$prefix.internStrings", Some(false))
 
-  // this class pre-processes the text before any computation happens
-  lazy val tokenizerPreProcessor:Option[TokenizerPreProcessor] =
-    getArgString(s"$prefix.tokenizer.pre.type", Some("none")) match {
-      case "bio" => Some(new BioTokenizerPreProcessor(
-        removeFigTabReferences = getArgBoolean(s"$prefix.tokenizer.pre.removeFigTabReferences", Some(true)),
-        removeBibReferences = getArgBoolean(s"$prefix.tokenizer.pre.removeBibReferences", Some(true))))
-      case "none" => None
-      case _ => throw new RuntimeException(s"ERROR: Unknown argument value for $prefix.tokenizer.pre.type!")
-    }
-
-  // the actual tokenizer
-  lazy val tokenizer: Tokenizer =
-    new OpenDomainEnglishTokenizer
-
   // this class post-processes the tokens produced by the tokenizer
-  lazy val tokenizerPostProcessor:Option[TokenizerPostProcessor] =
+  lazy private val tokenizerPostProcessor:Option[TokenizerStep] =
     getArgString(s"$prefix.tokenizer.post.type", Some("none")) match {
       case "bio" => Some(new BioTokenizerPostProcessor(
         getArgStrings(s"$prefix.tokenizer.post.tokensWithValidSlashes", None)
@@ -58,6 +39,14 @@ class CluProcessor (val config: Config = ConfigFactory.load("cluprocessoropen"))
       case "none" => None
       case _ => throw new RuntimeException(s"ERROR: Unknown argument value for $prefix.tokenizer.post.type!")
     }
+
+  // the actual tokenizer
+  lazy val tokenizer: Tokenizer =
+    new OpenDomainEnglishTokenizer(tokenizerPostProcessor)
+
+  // the lemmatizer
+  lazy val lemmatizer: Lemmatizer =
+    new EnglishLemmatizer
 
   // the POS tagger
   lazy val posTagger: PartOfSpeechTagger =
@@ -125,72 +114,16 @@ class CluProcessor (val config: Config = ConfigFactory.load("cluprocessoropen"))
     doc
   }
 
-  override def preprocessText(origText:String):String = {
-    if(tokenizerPreProcessor.nonEmpty) tokenizerPreProcessor.get.process(origText)
-    else origText
-  }
-
-  private def postProcess(document: Document): Document = {
-    if(tokenizerPostProcessor.isEmpty) return document
-    val sentences = new ArrayBuffer[Sentence]()
-    for(s <- document.sentences) {
-      // this method is called before any real NLP happens, so POS tags, NER, etc. should be empty
-      assert(s.tags.isEmpty)
-      assert(s.lemmas.isEmpty)
-      assert(s.entities.isEmpty)
-
-      val tokens = mkPostProcessorTokens(s)
-      val newTokens = tokenizerPostProcessor.get.process(tokens)
-      sentences += mkSentence(newTokens)
-    }
-
-    // this method is called before any real NLP happens, so coref and discourse should be empty
-    assert(document.coreferenceChains.isEmpty)
-    assert(document.discourseTree.isEmpty)
-    val d = new Document(sentences.toArray)
-    d.text = document.text
-    d
-  }
-
-  private def mkPostProcessorTokens(sentence: Sentence):Array[PostProcessorToken] = {
-    val tokens = new Array[PostProcessorToken](sentence.size)
-    for(i <- sentence.indices) {
-      tokens(i) = PostProcessorToken(sentence.words(i), sentence.startOffsets(i), sentence.endOffsets(i))
-    }
-    tokens
-  }
-
-  private def mkSentence(tokens:Array[PostProcessorToken]): Sentence = {
-    val words = new Array[String](tokens.length)
-    val startOffsets = new Array[Int](tokens.length)
-    val endOffsets = new Array[Int](tokens.length)
-    for(i <- tokens.indices) {
-      words(i) = tokens(i).word
-      startOffsets(i) = tokens(i).beginPosition
-      endOffsets(i) = tokens(i).endPosition
-    }
-    new Sentence(words, startOffsets, endOffsets)
-  }
-
   /** Constructs a document of tokens from free text; includes sentence splitting and tokenization */
   def mkDocument(text:String, keepText:Boolean = false): Document = {
-    val sents = tokenizer.tokenize(text)
-    val doc = new Document(sents)
-    if(keepText) doc.text = Some(text)
-    postProcess(doc)
+    CluProcessor.mkDocument(tokenizer, text, keepText)
   }
 
   /** Constructs a document of tokens from an array of untokenized sentences */
   def mkDocumentFromSentences(sentences:Iterable[String],
                               keepText:Boolean = false,
                               charactersBetweenSentences:Int = 1): Document = {
-    val sents = new ArrayBuffer[Sentence]()
-    for(text <- sentences) {
-      sents ++= tokenizer.tokenize(text, sentenceSplit = false) // we produce a single sentence here!
-    }
-    val doc = new Document(sents.toArray)
-    if(keepText) doc.text = Some(sentences.mkString(mkSep(charactersBetweenSentences)))
-    postProcess(doc)
+    CluProcessor.mkDocumentFromSentences(tokenizer, sentences, keepText, charactersBetweenSentences)
   }
 
   /** Constructs a document of tokens from an array of tokenized sentences */
@@ -198,40 +131,7 @@ class CluProcessor (val config: Config = ConfigFactory.load("cluprocessoropen"))
                            keepText:Boolean = false,
                            charactersBetweenSentences:Int = 1,
                            charactersBetweenTokens:Int = 1): Document = {
-    var charOffset = 0
-    var sents = new ArrayBuffer[Sentence]()
-    val text = new StringBuilder
-    for(sentence <- sentences) {
-      val startOffsets = new ArrayBuffer[Int]()
-      val endOffsets = new ArrayBuffer[Int]()
-      for(word <- sentence) {
-        startOffsets += charOffset
-        charOffset += word.length
-        endOffsets += charOffset
-        charOffset += charactersBetweenTokens
-      }
-      sents += new Sentence(sentence.toArray, startOffsets.toArray, endOffsets.toArray)
-      charOffset += charactersBetweenSentences
-      if(keepText) {
-        text.append(sentence.mkString(mkSep(charactersBetweenTokens)))
-        text.append(mkSep(charactersBetweenSentences))
-      }
-    }
-
-    val doc = new Document(sents.toArray)
-    if(keepText) doc.text = Some(text.toString)
-    postProcess(doc)
-  }
-
-  private def mkSep(size:Int):String = {
-    val os = new mutable.StringBuilder
-    for (_ <- 0 until size) os.append(" ")
-    os.toString()
-  }
-
-  private def in(s:String):String = {
-    if (internStrings) Processor.internString(s)
-    else s
+    CluProcessor.mkDocumentFromTokens(tokenizer, sentences, keepText, charactersBetweenSentences, charactersBetweenTokens)
   }
 
   /** Part of speech tagging */
@@ -247,29 +147,6 @@ class CluProcessor (val config: Config = ConfigFactory.load("cluprocessoropen"))
     }
   }
 
-  def lemmatizeWord(word:String):String = {
-    val norm = CluProcessor.normalizeForLemmatization(word).trim
-    if(norm.isEmpty) return ""
-    
-    val parts = norm.split(whitespaces)
-
-    val result = new mutable.StringBuilder()
-    for(part <- parts) {
-      val morpha = new Morpha(new StringReader(part), false)
-
-      var lemma = part
-      try {
-        lemma = morpha.next()
-      } catch {
-        case _:Throwable =>
-      }
-
-      if(result.length > 0) result.append(" ")
-      result.append(lemma)
-    }
-    result.toString()
-  }
-
   /** Lematization; modifies the document in place */
   def lemmatize(doc:Document) {
     basicSanityCheck(doc)
@@ -277,11 +154,8 @@ class CluProcessor (val config: Config = ConfigFactory.load("cluprocessoropen"))
       //println(s"Lemmatize sentence: ${sent.words.mkString(", ")}")
       val lemmas = new Array[String](sent.size)
       for(i <- sent.words.indices) {
-        var lemma = lemmatizeWord(sent.words(i))
-        // in some case, Morpha returns empty strings
-        if(lemma.isEmpty)
-          lemma = sent.words(i).toLowerCase()
-        lemmas(i) = lemma
+        lemmas(i) = lemmatizer.lemmatizeWord(sent.words(i))
+        assert(lemmas(i).nonEmpty)
       }
       sent.lemmas = Some(lemmas)
     }
@@ -312,6 +186,8 @@ class CluProcessor (val config: Config = ConfigFactory.load("cluprocessoropen"))
 
     for (sentence <- doc.sentences) {
       //println(s"PARSING SENTENCE: ${sentence.words.mkString(", ")}")
+      //println(sentence.tags.get.mkString(", "))
+      //println(sentence.lemmas.get.mkString(", "))
       val dg = depParser.parseSentence(sentence)
 
       if(useUniversalDependencies) {
@@ -354,14 +230,6 @@ class CluProcessor (val config: Config = ConfigFactory.load("cluprocessoropen"))
 
 }
 
-trait TokenizerPreProcessor {
-  def process(text:String):String
-}
-
-trait TokenizerPostProcessor {
-  def process(tokens:Array[PostProcessorToken]):Array[PostProcessorToken]
-}
-
 trait SentencePostProcessor {
   def process(sentence: Sentence)
 }
@@ -374,14 +242,82 @@ object CluProcessor {
   val logger:Logger = LoggerFactory.getLogger(classOf[CluProcessor])
   val prefix:String = "CluProcessor"
 
-  /** Special characters to remove. */
-  val remove: Regex = """[()\[\].,;:"']""".r
+  /** Constructs a document of tokens from free text; includes sentence splitting and tokenization */
+  def mkDocument(tokenizer:Tokenizer,
+                 text:String,
+                 keepText:Boolean): Document = {
+    val sents = tokenizer.tokenize(text)
+    val doc = new Document(sents)
+    if(keepText) doc.text = Some(text)
+    doc
+  }
 
-  /** White spaces */
-  val whitespaces: String = "\\s+"
+  /** Constructs a document of tokens from an array of untokenized sentences */
+  def mkDocumentFromSentences(tokenizer:Tokenizer,
+                              sentences:Iterable[String],
+                              keepText:Boolean,
+                              charactersBetweenSentences:Int): Document = {
+    val sents = new ArrayBuffer[Sentence]()
+    var characterOffset = 0
+    for(text <- sentences) {
+      val sent = tokenizer.tokenize(text, sentenceSplit = false).head // we produce a single sentence here!
 
-  /** Remove special characters and lowercase the string. */
-  def normalizeForLemmatization(word: String):String = CluProcessor.remove.replaceAllIn(word.trim.toLowerCase, "")
+      // update character offsets between sentences
+      for(i <- 0 until sent.size) {
+        sent.startOffsets(i) += characterOffset
+        sent.endOffsets(i) += characterOffset
+      }
+
+      // move the character offset after the current sentence
+      characterOffset = sent.endOffsets.last + charactersBetweenSentences
+
+      //println("SENTENCE: " + sent.words.mkString(", "))
+      //println("Start offsets: " + sent.startOffsets.mkString(", "))
+      //println("End offsets: " + sent.endOffsets.mkString(", "))
+      sents += sent
+    }
+    val doc = new Document(sents.toArray)
+    if(keepText) doc.text = Some(sentences.mkString(mkSep(charactersBetweenSentences)))
+    doc
+  }
+
+  /** Constructs a document of tokens from an array of tokenized sentences */
+  def mkDocumentFromTokens(tokenizer:Tokenizer,
+                           sentences:Iterable[Iterable[String]],
+                           keepText:Boolean,
+                           charactersBetweenSentences:Int,
+                           charactersBetweenTokens:Int): Document = {
+    var charOffset = 0
+    var sents = new ArrayBuffer[Sentence]()
+    val text = new StringBuilder
+    for(sentence <- sentences) {
+      val startOffsets = new ArrayBuffer[Int]()
+      val endOffsets = new ArrayBuffer[Int]()
+      for(word <- sentence) {
+        startOffsets += charOffset
+        charOffset += word.length
+        endOffsets += charOffset
+        charOffset += charactersBetweenTokens
+      }
+      // note: NO postprocessing happens in this case, so use it carefully!
+      sents += new Sentence(sentence.toArray, startOffsets.toArray, endOffsets.toArray, sentence.toArray)
+      charOffset += charactersBetweenSentences - charactersBetweenTokens
+      if(keepText) {
+        text.append(sentence.mkString(mkSep(charactersBetweenTokens)))
+        text.append(mkSep(charactersBetweenSentences))
+      }
+    }
+
+    val doc = new Document(sents.toArray)
+    if(keepText) doc.text = Some(text.toString)
+    doc
+  }
+
+  private def mkSep(size:Int):String = {
+    val os = new mutable.StringBuilder
+    for (_ <- 0 until size) os.append(" ")
+    os.toString()
+  }
 }